+++ haskell/01/ex1-38.hs
@@ -3,15 +3,15 @@
 -- e = 2.718..
 import ContFrac
 
-d :: Int -> Float
+d :: Int -> Double
 d k
     | mod k 3 == 2  = 2.0 * fromIntegral (1 + (k `div` 3))
     | otherwise     = 1.0
 
-n :: Int -> Float
+n :: Int -> Double
 n = const 1
 
-euler :: Int -> Float
+euler :: Int -> Double
 euler k = 2.0 + contFrac n d k
 
 main = do
