+++ haskell/01/ex1-39.hs
@@ -3,14 +3,14 @@
 -- ランベルトの式
 import ContFrac
 
-d :: Int -> Float
+d :: Int -> Double
 d k = fromIntegral k * 2 - 1
 
-n :: Float -> Int -> Float
+n :: Double -> Int -> Double
 n x 1 = x
 n x _ = - x ^ 2
 
-tanCf :: Float -> Int -> Float
+tanCf :: Double -> Int -> Double
 tanCf x = contFrac (n x) d
 
 main = do
